Move 2 commands into the core directory, and thus into the core package in help.
