+++ lib/Drush/CommandFiles/core/EvalCommandFile.php
@@ -1,5 +1,5 @@
 <?php
-namespace Drush\CommandFiles;
+namespace Drush\CommandFiles\core;
 
 /**
  * @file
